MQTT moved to separate class but it doesn't work. Line 104 doesn't work and I don't know how to fix it.
