+++ android/app/src/main/java/com/example/android/MainActivity.java
@@ -21,6 +21,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
     @Override
     public void onClick(View v) {
+
         Intent intent = new Intent(this, ManualActivity.class);
         startActivity(intent);
     }
